correcting tests

+++ src/test-c/test-ok/eg_db_t.c
@@ -1,5 +1,5 @@
-double main(){
+int main(){
     double i;
     i= (1.2==1.2);
-    return i-1;
+    return (int)i-1;
 }
