+++ src/test-c/test-ok/int_db_t.c
@@ -1,7 +1,7 @@
-double main(){
+int main(){
     double i,j,k;
     j = 1;
     k = 2;
     i = (j<k);
-    return i-1;
+    return (int) i-1;
 }
